ver2

+++ main.cpp
@@ -4,8 +4,8 @@ int main()
 {
 	// EX Begin8
 	
-	unsigned int a, b, Am; // variables
-	
+	double a, b, Am; // variables
+
 	cout << "Enter a: "; // introduct var 'a'
 	cin >> a;
 	cout << "Enter b: "; // introduct var 'b'
@@ -26,7 +26,7 @@ int main()
 	cout << "Enter T (hours): "; // introduct var 'T' - time
 	cin >> T;
 
-	S= V1*T + V2*T; // result calculation total distance
+	S= V1*T + V2*T; // total distance calculation
 
 	cout << "Total distance between cars: " << S << endl; // result output
